+++ src/commands/size.js
@@ -20,7 +20,7 @@ const Checks = require("../classes/Checks");
  * @param {string} key The key, which is the name of the type of size.
  * @param {number | string} value The value of the type of size in `key`.
  */
-const _logSizes = (key, value) => {
+const _logSize = (key, value) => {
   // Capitalize the first letter of the word
   const showSize = key.charAt(0).toUpperCase() + key.slice(1);
 
@@ -33,7 +33,31 @@ const _logSizes = (key, value) => {
   }
 };
 
-const size = (file, sizes, ...args) => {
+/**
+ * Show the size of a file from the BubbleOS CLI.
+ * This is only meant to be used inside of the
+ * BubbleOS shell and should not be used elsewhere.
+ *
+ * Usage:
+ *
+ * ```js
+ * size("test.txt"); // Arguments accepted!
+ * ```
+ *
+ * Note that getting the size of a directory is not
+ * yet supported, however, this is to be added in
+ * BubbleOS build 200.
+ *
+ * Available arguments:
+ * - `-b`: Get the bytes of a file only.
+ * - `-kb`: Get the megabytes of a file only.
+ * - `-mb`: Get the kilobytes of a file only.
+ * - `-gb`: Get the gigabytes of a file only.
+ *
+ * @param {fs.PathLike | string} file The file to find the sizes of.
+ * @param  {...string} args The arguments to change the behaviour of `size`.
+ */
+const size = (file, ...args) => {
   try {
     // Replace spaces and convert it to an absolute path
     file = _convertAbsolute(_replaceSpaces(file));
@@ -41,42 +65,20 @@ const size = (file, sizes, ...args) => {
     // Initialize checker
     const fileChk = new Checks(file);
 
+    // Initialize arguments
+    // Sizes
+    const bytes = args?.includes("-b") || args?.includes("/b");
+    const kilobytes = args?.includes("-kb") || args?.includes("/kb");
+    const megabytes = args?.includes("-mb") || args?.includes("/mb");
+    const gigabytes = args?.includes("-gb") || args?.includes("/gb");
+
+    // If no arguments were passed, show all
+    const all = !bytes && !kilobytes && !megabytes && !gigabytes;
+
     // If the file is not defined
     if (fileChk.paramUndefined()) {
       Errors.enterParameter("a file", "size test.txt");
       return;
-    }
-
-    // Sizes translations for the 'sizes' parameter
-    const sizesTrans = {
-      b: "Bytes",
-      kb: "Kilobytes",
-      mb: "Megabytes",
-      gb: "Gigabytes",
-      a: "All",
-    };
-
-    // If the 'sizes' is defined...
-    if (!new Checks(sizes).paramUndefined()) {
-      // Split it by commas
-      sizes = sizes.split(",");
-
-      // If the size the user requested is in the translation, leave it, else, delete it
-      for (const val in sizesTrans) {
-        if (!sizes.includes(val)) {
-          delete sizesTrans[val];
-        }
-      }
-
-      // If there are no more translations left
-      if (Object.keys(sizesTrans).length === 0) {
-        console.log(
-          chalk.yellow(
-            `No matches for the filter ${chalk.bold.italic(`'${sizes}'`)}.\nOperation cancelled.\n`
-          )
-        );
-        return;
-      }
     }
 
     if (!fileChk.doesExist()) {
@@ -92,26 +94,11 @@ const size = (file, sizes, ...args) => {
     // The size shortened to 4 decimal places
     const allSizes = _convertSize(fs.statSync(file).size, 4);
 
-    // If the sizes were passed in
-    if (typeof sizes !== "undefined") {
-      // If the user requested to see all ('a')
-      if (typeof sizesTrans.a !== "undefined") {
-        // Show all values
-        for (const size in allSizes) {
-          _logSizes(size, allSizes[size]);
-        }
-      } else {
-        // Only show the values that the user requested
-        for (const val in sizesTrans) {
-          _logSizes(sizesTrans[val], allSizes[sizesTrans[val].toLowerCase()]);
-        }
-      }
-    } else {
-      // Show all of the values
-      for (const size in allSizes) {
-        _logSizes(size, allSizes[size]);
-      }
-    }
+    // Log all sizes depending on what sizes the user requested
+    if (bytes || all) _logSize("Bytes", allSizes.bytes);
+    if (kilobytes || all) _logSize("Kilobytes", allSizes.kilobytes);
+    if (megabytes || all) _logSize("Megabytes", allSizes.megabytes);
+    if (gigabytes || all) _logSize("Gigabytes", allSizes.gigabytes);
 
     // Newline and return
     console.log();